refactor: remove unused code and fix import paths
- Remove duplicate UI components (Button, Input) and unused components (Modal, ProgressiveImage, Dialog, Spinner)
- Remove unused utilities (analytics, validation) - ~600+ lines of unused code
- Remove unused hooks (usePerformance)
- Fix import paths in movie pages to use new feature-based architecture
- Update index files to remove exports for deleted components
- Maintain all existing functionality with cleaner, more maintainable codebase
