--- src/components/ui/Button/Button.tsx
@@ -1,77 +1,0 @@
-import React from 'react';
-import { tv, type VariantProps } from 'tailwind-variants';
-
-const buttonVariants = tv({
-  base: 'inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50',
-  variants: {
-    variant: {
-      default: 'bg-secondary-500 text-white hover:bg-secondary-600',
-      destructive: 'bg-red-500 text-white hover:bg-red-600',
-      outline: 'border border-gray-300 bg-white hover:bg-gray-50 hover:text-gray-900',
-      secondary: 'bg-gray-200 text-gray-900 hover:bg-gray-300',
-      ghost: 'hover:bg-gray-100 hover:text-gray-900',
-      link: 'text-secondary-600 underline-offset-4 hover:underline',
-      tertiary: 'bg-tertiary-500 text-white hover:bg-tertiary-600',
-    },
-    size: {
-      default: 'h-10 px-4 py-2',
-      sm: 'h-9 rounded-md px-3',
-      lg: 'h-11 rounded-md px-8',
-      icon: 'h-10 w-10',
-    },
-  },
-  defaultVariants: {
-    variant: 'default',
-    size: 'default',
-  },
-});
-
-export interface ButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof buttonVariants> {
-  asChild?: boolean;
-  loading?: boolean;
-}
-
-const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
-  ({ className, variant, size, asChild, loading = false, children, disabled, ...props }, ref) => {
-    // asChild is reserved for future implementation when we need to render as different elements
-    void asChild; // Mark as intentionally unused
-    return (
-      <button
-        className={buttonVariants({ variant, size, class: className })}
-        ref={ref}
-        disabled={disabled || loading}
-        {...props}
-      >
-        {loading && (
-          <svg
-            className="mr-2 h-4 w-4 animate-spin"
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 24 24"
-          >
-            <circle
-              className="opacity-25"
-              cx="12"
-              cy="12"
-              r="10"
-              stroke="currentColor"
-              strokeWidth="4"
-            ></circle>
-            <path
-              className="opacity-75"
-              fill="currentColor"
-              d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-            ></path>
-          </svg>
-        )}
-        {children}
-      </button>
-    );
-  }
-);
-
-Button.displayName = 'Button';
-
-export default Button;
