--- src/components/ui/Button/Button.types.ts
@@ -1,4 +1,0 @@
-export interface ButtonVariants {
-  variant?: 'default' | 'destructive' | 'outline' | 'secondary' | 'ghost' | 'link' | 'tertiary';
-  size?: 'default' | 'sm' | 'lg' | 'icon';
-}
