--- src/components/ui/Button/index.ts
@@ -1,3 +1,0 @@
-export { default as Button } from './Button';
-export type { ButtonProps } from './Button';
-export type { ButtonVariants } from './Button.types';
