--- src/components/ui/Input/Input.tsx
@@ -1,101 +1,0 @@
-import React, { forwardRef, useId } from 'react';
-import { cn } from '@utils/cn';
-
-export interface InputProps
-  extends React.InputHTMLAttributes<HTMLInputElement> {
-  label?: string;
-  error?: string;
-  helperText?: string;
-  startIcon?: React.ReactNode;
-  endIcon?: React.ReactNode;
-  containerClassName?: string;
-}
-
-const Input = forwardRef<HTMLInputElement, InputProps>(
-  (
-    {
-      className,
-      type,
-      label,
-      error,
-      helperText,
-      startIcon,
-      endIcon,
-      containerClassName,
-      id,
-      ...props
-    },
-    ref
-  ) => {
-    const generatedId = useId();
-    const inputId = id || `input-${generatedId}`;
-
-    return (
-      <div className={cn('space-y-2', containerClassName)}>
-        {label && (
-          <label
-            htmlFor={inputId}
-            className="text-sm font-medium text-gray-700 leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70"
-          >
-            {label}
-          </label>
-        )}
-
-        <div className="relative">
-          {startIcon && (
-            <div className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400">
-              {startIcon}
-            </div>
-          )}
-
-          <input
-            type={type}
-            id={inputId}
-            className={cn(
-              'flex h-10 w-full rounded-md border border-gray-300 bg-white px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-gray-500 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-secondary-500 focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50',
-              startIcon && 'pl-10',
-              endIcon && 'pr-10',
-              error && 'border-red-500 focus-visible:ring-red-500',
-              className
-            )}
-            ref={ref}
-            {...props}
-          />
-
-          {endIcon && (
-            <div className="absolute right-3 top-1/2 transform -translate-y-1/2 text-gray-400">
-              {endIcon}
-            </div>
-          )}
-        </div>
-
-        {error && (
-          <p className="text-sm text-red-600 flex items-center gap-1">
-            <svg
-              className="h-4 w-4"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
-              />
-            </svg>
-            {error}
-          </p>
-        )}
-
-        {helperText && !error && (
-          <p className="text-sm text-gray-500">{helperText}</p>
-        )}
-      </div>
-    );
-  }
-);
-
-Input.displayName = 'Input';
-
-export default Input;
