--- src/components/ui/Input/index.ts
@@ -1,2 +1,0 @@
-export { default as Input } from './Input';
-export type { InputProps } from './Input';
