--- src/components/ui/Modal/Modal.tsx
@@ -1,96 +1,0 @@
-import React from 'react';
-import {
-  Dialog,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog';
-import { Button } from '@/components/ui/button';
-
-interface ModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  title?: string;
-  children: React.ReactNode;
-  size?: 'sm' | 'md' | 'lg' | 'xl' | 'full';
-  showCloseButton?: boolean;
-  closeOnOverlayClick?: boolean;
-  footer?: React.ReactNode;
-  className?: string;
-}
-
-const Modal: React.FC<ModalProps> = ({
-  isOpen,
-  onClose,
-  title,
-  children,
-  size = 'md',
-  showCloseButton = true,
-  closeOnOverlayClick = true,
-  footer,
-  className,
-}) => {
-  const sizeClasses = {
-    sm: 'max-w-md',
-    md: 'max-w-lg',
-    lg: 'max-w-2xl',
-    xl: 'max-w-4xl',
-    full: 'max-w-full mx-4',
-  };
-
-  return (
-    <Dialog open={isOpen} onOpenChange={(open) => !open && onClose()}>
-      <DialogContent
-        className={`${sizeClasses[size]} ${className || ''}`}
-        onPointerDownOutside={(e) => {
-          if (!closeOnOverlayClick) {
-            e.preventDefault();
-          }
-        }}
-      >
-        {(title || showCloseButton) && (
-          <DialogHeader>
-            {title && <DialogTitle>{title}</DialogTitle>}
-            {showCloseButton && (
-              <Button
-                onClick={onClose}
-                variant="ghost"
-                size="icon"
-                className="h-8 w-8 rounded-full p-0 hover:bg-accent"
-                aria-label="Close modal"
-              >
-                <svg
-                  className="h-4 w-4"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M6 18L18 6M6 6l12 12"
-                  />
-                </svg>
-              </Button>
-            )}
-          </DialogHeader>
-        )}
-
-        <div className="py-4">
-          {children}
-        </div>
-
-        {footer && (
-          <DialogFooter>
-            {footer}
-          </DialogFooter>
-        )}
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-export { Modal };
-export type { ModalProps };
