--- src/components/ui/Modal/index.ts
@@ -1,2 +1,0 @@
-export { Modal } from './Modal';
-export type { ModalProps } from './Modal';
