--- src/components/ui/ProgressiveImage/index.ts
@@ -1,2 +1,0 @@
-export { ProgressiveImage } from './ProgressiveImage';
-export type { ProgressiveImageProps } from './ProgressiveImage';
