+++ src/components/ui/index.ts
@@ -4,27 +4,9 @@ export { Button } from './button';
 export { Input } from './input';
 
 export {
-  Dialog,
-  DialogPortal,
-  DialogOverlay,
-  DialogTrigger,
-  DialogClose,
-  DialogContent,
-  DialogHeader,
-  DialogFooter,
-  DialogTitle,
-  DialogDescription
-} from './dialog';
-
-export {
   Alert,
   AlertDescription,
   AlertTitle
 } from './alert';
 
-export { Spinner } from './spinner';
-
 export { Skeleton } from './skeleton';
-
-// Legacy exports for backward compatibility
-export { Button as CustomButton } from './Button';
