+++ src/utils/index.ts
@@ -3,4 +3,3 @@ export { envConfig } from './env';
 export { withApiKey, withAuth, withoutAuth } from './axios';
 export * from './constants';
 export * from './formatting';
-export * from './validation';
